refactor: admin, home, myproducts, category, detailedProduct, checkout refactoring
