+++ my-monorepo/apps/shopease/src/pages/administration/components/AdminTitle.tsx
@@ -1,0 +1,12 @@
+const AdminTitle = () => {
+  return (
+    <div className="flex flex-col items-center justify-center w-full h-48 border-b-[1px] border-gray-light">
+      <h1 className="text-primary font-semibold text-[50px]">Administration</h1>
+      <p>
+        관리자 페이지에선 주문 완료된 상품들을 관리할 수 있는 페이지 입니다.
+      </p>
+    </div>
+  );
+};
+
+export default AdminTitle;
