+++ my-monorepo/apps/shopease/src/pages/categoryProduct/index.tsx
@@ -4,22 +4,21 @@ import NavigationBar from "../../shared/layout/NavigationBar";
 import CategoryFilterButton from "./components/CategoryFilterButton";
 import React, { Suspense } from "react";
 import { LoadingSkeleton } from "../../shared/layout/LoadingSkeleton";
-import GetCategoryTitle from "./components/GetCategoryTitle";
+import useCategoryTitle from "./components/useCategoryTitle";
 
 const CategoryCard = React.lazy(() => import("./components/CategoryCard"));
 
 const CategoryProduct = () => {
+  const categoryTitle = useCategoryTitle();
   return (
     <Layout>
-      <div>
-        <NavigationBar />
-        <HomeButton style="absolute top-32 left-10" />
-        <div className="flex items-center justify-center w-full h-48 border-b-[1px] border-gray-light">
-          <h1 className="text-primary font-semibold text-[50px]">Categories</h1>
-        </div>
+      <NavigationBar />
+      <HomeButton style="absolute top-32 left-10" />
+      <div className="flex items-center justify-center w-full h-48 border-b-[1px] border-gray-light">
+        <h1 className="text-primary font-semibold text-[50px]">Categories</h1>
       </div>
       <div className="mx-14 my-10 text-4xl font-semibold text-primary">
-        <GetCategoryTitle />
+        {categoryTitle}
       </div>
       <CategoryFilterButton />
       <Suspense
