+++ my-monorepo/apps/shopease/src/pages/myProducts/index.tsx
@@ -2,7 +2,7 @@ import { LoadingSkeleton } from "@/shared/layout/LoadingSkeleton";
 import HomeButton from "../../shared/layout/HomeButton";
 import { Layout, authStatusType } from "../../shared/layout/Layout";
 import NavigationBar from "../../shared/layout/NavigationBar";
-import RegisterModal from "./components/RegisterModal";
+import MyProductsRegisterModal from "./components/MyProductsRegisterModal";
 import React, { Suspense } from "react";
 
 const MyProductsCard = React.lazy(() => import("./components/MyProductsCard"));
@@ -10,15 +10,12 @@ const MyProductsCard = React.lazy(() => import("./components/MyProductsCard"));
 const MyProducts = () => {
   return (
     <Layout authStatus={authStatusType.SELLER}>
-      <div>
-        <NavigationBar />
-        <HomeButton style="absolute top-32 left-10" />
-        <div className="flex items-center justify-center w-full h-48 border-b-[1px] border-gray-light">
-          <h1 className="text-primary font-semibold text-[50px]">MyProducts</h1>
-        </div>
+      <NavigationBar />
+      <HomeButton style="absolute top-32 left-10" />
+      <div className="flex items-center justify-center w-full h-48 border-b-[1px] border-gray-light">
+        <h1 className="text-primary font-semibold text-[50px]">MyProducts</h1>
       </div>
-
-      <RegisterModal />
+      <MyProductsRegisterModal />
       <Suspense
         fallback={
           <LoadingSkeleton
